TH: Update battery consumption computation to include charging behavior

+++ src/auto_smart_factory/include/agent/task_handling/TaskHandler.h
@@ -45,7 +45,7 @@ class TaskHandler
 
 		Task* getCurrentTask(void);
 
-		float getBatteryConsumption(void);
+		double getBatteryConsumption(void);
 
 		double getDuration(void);
